# Disable BLAS threading within `conv_im2col!` etc.