--- a/src/conv.jl
+++ b/src/conv.jl
@@ -196,11 +196,22 @@ for (front_name, backend) in (
                                  C_in = channels_in(cdims) ÷ groupcount(cdims),
                                  C_out = channels_out(cdims) ÷ groupcount(cdims))
 
-            Threads.@sync for (xc, wc) in zip(x_cs, w_cs)
+            if nthreads() > 1
+                th = BLAS.get_num_threads()
+                BLAS.set_num_threads(1)
+                # conv_im2col! has a loop with @threads, and benchmarks show that this is usually
+                # faster without BLAS multithreading, and without @spawn in the zip(x_cs, w_cs) loop.
+            end
+
+            for (xc, wc) in zip(x_cs, w_cs)
                 x = @view in1[ntuple(i -> i == 4 ? xc : Colon(), 5)...]
                 w = @view in2[ntuple(i -> i == 5 ? wc : Colon(), 5)...]
                 y = @view out[ntuple(i -> i == 4 ? wc : Colon(), 5)...]
-                Threads.@spawn $(Symbol("$(front_name)_$(backend)!"))(y, x, w, cdims2; kwargs...)
+                $(Symbol("$(front_name)_$(backend)!"))(y, x, w, cdims2; kwargs...)
+            end
+
+            if nthreads() > 1
+                BLAS.set_num_threads(th)
             end
 
             return out