@darsnack 2022-02-22
Shouldn't this also test a convolution with groups > 1? That's what the spawn was for.