@mcabbott 2022-02-22
Maybe, got an example?

In this case the `@spawn` is pure loss of course. Ideally I think we want exactly one multi-threading loop, but where to put it (without too much complexity) seems messy. Maybe making the batch loop use `@spawn` might make them compose better.