@CarloLucibello 2022-04-02
Shall we go on with this? It could test also the grouped conv in https://github.com/FluxML/Flux.jl/pull/1921#issuecomment-1086576955

I can benchmark on a amd threadripper platform