@mcabbott 2022-04-02
Benchmarks would be great. I stopped here because I thought we ought to check a whole range -- what if there is one group, or fewer channels than threads, etc... just timing one thing could lead you anywhere. 

Ideally we'd have multi-threading only on one outermost loop (and with some heuristic to decide when the problem is big enough). If one loop can't cover all cases then IIRC `@spawn` within `@spawn` composes better than `@threads`. But sorting all of this out started to seem like a big project.