@ToucheSir 2022-04-07
@CarloLucibello thoughts on expanding https://github.com/FluxML/Flux.jl/blob/master/perf/conv.jl to use as a benchmarking suite?